requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
culqipy==0.2.6
Django==1.11.6
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
Markdown==2.6.8
psycopg2==2.7.1
pytz==2017.2
requests==2.18.4
urllib3==1.22
vine==1.1.4
